Scroll to the channel when changing the horizontal channel
Previously, when running CTRL+N or CTRL+P to go to the next or previous
channel, with the horizontal channel mode, the selected channel could
appear offscreen.

This patch ensures that we scroll just enough to make the newly selected
chan visible.

Thanks to Nomeji for his hard work on this. Adapted slightly to
refactor existing code to use the newly created width computation
routine.