Revert to using base colors by default, making extended configurable
In a previous commit, the color scheme for displaying nicks was changed
to use more colors (30 rather than 14), using extended colors from the
256 colors set.

The issue with this is that most terminal emulators customize the colors
of the first 16 colors to make them more readable (eg, a terminal
emulator with a light theme will make colors darker, and one with a dark
theme will make colors lighter). So the 14 colors used before were
usually not the default color codes from the original xterm/X11 spec,
but specific colors chosen by the terminal emulator to be particularly
readable.

In a way, changing the behavior to use colors from the 256 colors set,
which is usually not overriden, made them much less readable.

Which is why we need a configuration option for this.

I conversatively chose to make the default color scheme the previous
one, with only the base 16 colors.

The 256 colors scheme can be enable by adding to the configuration file:

  colors {
    nicks extended
  }