Highlight only on word boundaries
Preivously, we highlighted the user for any message that contained the
nick, even in the middle of a word.

Now, we only send a highlight if the user was mentioned in a single
word, by detecting word boundaries around the mention of the user.

For example, considering user "foo":

Previously, we would highlight on "foo" and "foolish".
Now, we highlight only on "foo".

This was inspired/copied from soju's handling [1].

[1]: https://git.sr.ht/~emersion/soju/tree/5e56cc30c538eb88c21cce732e2e357e80098164/item/irc.go#L473

Fixes: https://todo.sr.ht/~taiite/senpai/97