Fix requesting CHATHISTORY TARGETS with invalid timestamps
Previously, we sent the CHATHISTORY TARGETS start time with an incorrect
offset equal to the system time offset.

This means that we only reopened private buffers from the last app close
+ N hours on a system with a N-hour time offset.

This fixes the issue by saving the stamp as UTC, and also making sure to
format any time offset correctly.