Fix processing CHATHISTORY TAGMSG as standalone messages
Receiving a TAGMSG in chathistory causes the handler to return nil.

Previously this caused the handler to be called again without the
playblack flag, causing us to emit events out of their history context.

This meant we would show typing notifications for users from chathistory
batches.

Soju will soon be able to send TAGMSG messages in its chathistory, so
we're fixing that now.