Process "last seen" markers of joined channels
Previously, senpai received MARKREAD events of newly joined channels but
dropped them, because they were received before the RPL_ENDOFNAMES which
marks the ends of the initial channel join batch.

This fixes the issue by storing the read marker, then using it in the
channel join event so that it can be set in the channel.