Show an "unread" indicator bar
When switching to a channel, show a ruler at the last unread position.

Keep the previous "last read" ruler for the focused buffer, since the
"last read" timestamp is updated as soon as the buffer is focused (which
would result in the ruler always rendering at the bottom).

Also add a special case in `SetRead` for when the "last read" timestamp
is received from the server _after_ the buffer was focused. This happens
basically for the buffer that was focused at start-up.

Fixes: https://todo.sr.ht/~taiite/senpai/76
Co-authored-by: delthas <delthas@dille.cc>