Add icon, custom notifications through local integrations
This adds custom notifications sent directly (bypassing the
terminal), that enables us to automatically dismiss them, add a
nice icon, and to automatically switch to the appropriate buffer
when clicked.

This can be disabled with `local-integrations false` in the config
if any issues are encountered. If you do, please open a ticket so
this can be investigated. (I'm aware this might be a bit
experimental.)

This also adds support for OSC 176 ([1]), which enables users to
use senpai as an "application", meaning it can be started from the
application launcher, pinned to a task bar, have a proper icon and
standalone window grouping, etc. This is at the time of writing
only supported on foot.

[1]: https://gist.github.com/delthas/d451e2cc1573bb2364839849c7117239