Support STATUSMSG-prefixed messages
Parse STATUSMSG prefixes in incoming messages, and push those messages
to the correct channel with the prefix indicated in the interface.

This also marks channel-context messages with a prefix.

Changes to ui.Line were added in order to support multiple styles
in the head.

Implements: https://todo.sr.ht/~delthas/senpai/183
Co-Authored-By: Taavi Väänänen <taavi@majava.org>